Repositoryディレクトリ整理
