+++ ComparePrices/Repository/Auth/FirebaseAuthRepository.swift
@@ -1,5 +1,5 @@
 //
-//  AuthRepository.swift
+//  FirebaseAuthRepository.swift
 //  ComparePrices
 //
 //  Created by ororo on 2021/05/03.
@@ -8,11 +8,6 @@
 import Combine
 import FirebaseAuth
 import Foundation
-
-protocol AuthRepository {
-    func signInAnonymously() -> Future<Void, Error>
-    func isAuthenticated() -> Bool
-}
 
 final class FirebaseAuthRepository: AuthRepository {
     func signInAnonymously() -> Future<Void, Error> {
@@ -29,17 +24,5 @@ final class FirebaseAuthRepository: AuthRepository {
     
     func isAuthenticated() -> Bool {
         Auth.auth().currentUser != nil
-    }
-}
-
-final class MockAuthRepository: AuthRepository {
-    func signInAnonymously() -> Future<Void, Error> {
-        Future { promise in
-            promise(.success(()))
-        }
-    }
-    
-    func isAuthenticated() -> Bool {
-        true
     }
 }
