+++ ComparePrices/Repository/CommodityPrice/MockCommodityPriceRepository.swift
@@ -1,26 +1,12 @@
 //
-//  CommodityPriceRepository .swift
+//  MockCommodityPriceRepository.swift
 //  ComparePrices
 //
-//  Created by Ryo Narisawa on 2021/03/07.
+//  Created by ororo on 2021/05/03.
 //
 
 import Combine
 import Foundation
-
-protocol CommodityPriceRepository {
-    func addCommodityPrice(_ purchaseResult: CommodityPrice) -> Future<Void, Error>
-    
-    func removeCommodityPrice(_ purchaseResult: CommodityPrice) -> Future<Void, Error>
-    
-    func getLowestCommodityPrice(_ commodityId: UUID) -> Future<CommodityPrice?, Error>
-    
-    func getAllCommodityPrices() -> Future<[CommodityPrice], Error>
-    
-    func getCommodityPrices(_ commodityId: UUID) -> Future<[CommodityPrice]?, Error>
-    
-    func observeCommodityPrices() -> AnyPublisher<[CommodityPrice], Error>
-}
 
 final class MockCommodityPriceRepository: CommodityPriceRepository {
     private class SingletonCommodities {
