+++ ComparePrices/Repository/Shop/MockShopRepository.swift
@@ -1,24 +1,12 @@
 //
-//  StoreRepository.swift
+//  MockShopRepository.swift
 //  ComparePrices
 //
-//  Created by Ryo Narisawa on 2021/03/07.
+//  Created by ororo on 2021/05/03.
 //
 
 import Combine
 import Foundation
-
-protocol ShopRepository {
-    func addShop(_ shop: Shop) -> Future<Void, Error>
-    
-    func removeShop(_ shop: Shop) -> Future<Void, Error>
-
-    func observeShops() -> AnyPublisher<[Shop], Error>
-    
-    func getShop(_ shopId: UUID) -> Future<Shop, Error>
-    
-    func getShops() -> Future<[Shop], Error>
-}
 
 final class MockShopRepository: ShopRepository {
     static let tairayaUUID = UUID()
